Crate liftof_cc

Modules

constants

threads

Functions

prefix_tof_cpu

send_default_calibration

Default function that starts calibration on all RBs with default values.

send_ltb_threshold_set

Function that sends the threshold to be set on all or specific LTBs

send_moni

Function that manages moni commands from ground

send_moni_response

Function that just replies to a moni command send to tofcpu

send_noi_calibration

Function that starts voltage calibration on a specific RB with the specified voltage level and extras (not implemented)

send_ping

Function that manages ping commands from ground

send_ping_response

Function that just replies to a ping command send to tofcpu

send_power

Power function that targets the component specified, no ID

send_power_id

Power function that targets the component specified with supplied ID

send_power_preamp

Power function that targets the component specified with supplied ID

send_power_response

Function that just replies to a moni command send to tofcpu

send_preamp_bias_set

Function that sends the threshold to be set on all or specific LTBs

send_run_start

Default function that starts run data taking on all RBs with default values.

send_run_stop

Default function that starts run data taking on all RBs with default values.

send_systemd_reboot

Function that send a restart command to RBs

send_timing_calibration

Function that starts timing calibration on a specific RB with the specified voltage level and extras (not implemented)

send_voltage_calibration

Function that starts voltage calibration on a specific RB with the specified voltage level and extras (not implemented)