liftof_cc
Function calibrate_tof

pub fn calibrate_tof(
    thread_control: Arc<Mutex<ThreadControl>>,
    rb_list: &Vec<ReadoutBoard>,
    show_progress: bool,
)

Run a full tof calibration - RBCalibration

The purpose of the RB calibration to is to create a relationship between the adc/timing bins and voltages and nanoseconds

This function is blocking, until a certain (configurable) timeout is expired. The timeout can be set in the configuration file

Argumeents:

thread_control : general shared memory to hold configuration settings, program st ate

rb_list : List of active readoutboards

show_progress : if true, it will show a progressbar with indicatif