liftof_cc
Function ssh_command_rbs

pub fn ssh_command_rbs(
    rb_list: &Vec<u8>,
    cmd: Vec<String>,
) -> Result<Vec<u8>, TofError>

Trigger a general command on the Readoutboards remotly through ssh.

Ssh keys and aliases (e.g. tof-rb02) must be set up for this to work

Arguments:

rb_list : The list of ReadoutBoard ids the commands will get executed

cmd : The actual command without ‘ssh ’

Returns:

A list of rb ids where the process failed