liftof_cc
Function verification_run

pub fn verification_run(
    timeout: u32,
    tp_to_sink: Sender<TofPacket>,
    thread_control: Arc<Mutex<ThreadControl>>,
)

A “verification” run describes an any trigger/track trigger run which should iluminate the entire tof so that we can do a working channel inventory.

A verification run will not safe data to disk, but instead run it through a small analysis engine and count the active channels