liftof_cc::threads

Module global_data_sink

Global data sink - a ‘funnel’ for all packets generated through the liftof system.

Each thread of liftof-cc can connect to the data sink through a channel and it will forward the tof packets to the designated zmq socket.

Functions

global_data_sink

Manages “outgoing” 0MQ PUB socket and writing data to disk