liftof_cc::threads::readoutboard_comm
Function readoutboard_communicator

pub fn readoutboard_communicator(
    ev_to_builder: Sender<RBEvent>,
    tp_to_sink: Sender<TofPacket>,
    rb: ReadoutBoard,
    thread_control: Arc<Mutex<ThreadControl>>,
)

Receive data from a readoutboard

In case of binary event data (“blob”) this can be analyzed here It is also possible to save the data directly.

In case of monitoring/other tof packets, those will be forwarded

Arguments:

ev_to_builder : This thread will receive RBEvent data from the assigned RB, if desired (see run_analysis_engine) run analysis and extract TofHits and then pass the result on to the event builder.

tp_to_sink : Channel which should be connect to a (global) data sink. Packets which are of not event type (e.g. header/full binary data) will be forwarded to the sink.

rb : ReadoutBoard instance, as loaded from the database. This will be used for readoutboard id as well as paddle assignment.

ae_settings : Settings to configure peakfinding algorithms etc. These can be configured with an external .toml file