Crate liftof_cc

Modules

api

API for liftof-cc, these are basically the individual threads

constants

threads

Functions

send_all_calibration

Default function that starts calibration on all RBs with default values.

send_timing_calibration

Function that starts timing calibration on a specific RB with the specified voltage level and extras (not implemented)

send_voltage_calibration

Function that starts voltage calibration on a specific RB with the specified voltage level and extras (not implemented)