Crate liftof_cc

Modules

constants

threads

Constants

LIFTOF_HOTWIRE

communicaton between liftof-scheduler and liftof-cc

Functions

calibrate_tof

Run a full tof calibration - RBCalibration

copy_file

copy_file_rename_liftof

delete_file

end_run

Regular run stop sequence

get_queue

Get the files in the queue and sort them by number

init_run_start

Regular run start sequence

manage_liftof_cc_service

Trigger a restart of liftof-cc and start a new run

move_file_rename_liftof

move_file_with_name

prepare_run

Prepare a new folder with the run id

rb_table

Produce a nicely formattable table with per RB information for scalar values

restart_liftof_rb

Restart liftof-rb on RBs

rm_liftof_rb_logs

Restart liftof-rb on RBs

run_cycler

Copy a config file from the queue to the current and next directories and restart liftof-cc.

set_rbs_sma_mode

Set SMA mode through rb-mode with tof-control’s binary

ssh_command_rbs

Trigger a general command on the Readoutboards remotly through ssh.