Module liftof_cc::constants

Constants

DEFAULT_CALIB_EXTRA

DEFAULT_CALIB_RB

DEFAULT_CALIB_VOLTAGE

! Command constants Calibration constants

EVENT_BUILDER_EVID_CACHE_SIZE

! Limit the size of the evids the event builder is currently waiting to get paddles for (this shoudl be rate*event_timeout

EVENT_CACHE_SIZE

! This should be rate dependent

EVENT_TIMEOUT

! How long to wait for paddles packets for each event in microseconds. This does highly depend on the frequency with which the readoutboards are emitting For now, lets use 30s. This will impact also the size of the caches (see below)

EXP_N_PADDLES_PER_EVENT

! Average number of paddle packets per event

MAX_NBOARDS

MAX_NUM_PEAKS

MAX_TRIGGER_RATE

! Expected maximum trigger rate

NCHN

NWORDS

PADDLE_PACKET_CACHE_SIZE

! Limit the size of the internal paddle packet cache