liftof_cc
Function prepare_run

pub fn prepare_run(
    data_path: String,
    config: &LiftofSettings,
    run_id: Option<u32>,
    create_dir: bool,
) -> Option<u32>

Prepare a new folder with the run id

This will assign a run id based on the run ids in that folder. Additionally, we will copy the current settings into that folder

TODO - connect it to the run database

Arguments:

data_path : The global path on the inflight tof computer where to store data.

config : The current configuration, to be copied into the new folder

run_id : Optionally define a pre-given run-id

create_dir : Create the directory for runfiles