Function liftof_cc::threads::command_dispatcher::command_dispatcher

pub fn command_dispatcher(
    settings: CommandDispatcherSettings,
    thread_ctrl: Arc<Mutex<ThreadControl>>,
    tof_ack_sender: Sender<TofPacket>,
    rb_ack_recv: Receiver<TofResponse>
)

The command dispatcher listens for incoming commands and either executes them or passes them on to the intended receiver

The acknowledgement packets will be just put into the general data stream and then be further processed by the receiver of that stream. This means, when we are taking data, they will be also logged to the disks on file

Arguments:

settings : Configure command_dispatcher with .toml config file

thread_ctrl : Interface with main program loop. E.g. shutdown, heartbeat signals.

tof_ack_sender : A channel in which we are putting the acknowledgement packets so that they can be further processed. This channel should connect to a data sink.

rb_ack_recv : Receive RB acknowledgements over this channel