Function liftof_cc::threads::event_builder::event_builder

pub fn event_builder(
    m_trig_ev: &Receiver<MasterTriggerEvent>,
    ev_from_rb: &Receiver<RBEvent>,
    data_sink: &Sender<TofPacket>
)

The event builder, assembling events from an id given by the master trigger

This requires the master trigger sending MasterTriggerEvents over the channel. The event builder then will querey the paddle_packet cache for paddle packets with the same event id Queries which can not be satisfied will lead to events being cached until they can be completed, or discarded after a timeout.

Arguments

m_trig_ev : Receive a MasterTriggerEvent over this channel. The event will be either build immediatly, or cached.

dsi_j_mapping : A HashMap of some Hashmaps which makes the following connection DSI/J/LTB_CH -> RBID, RBCH

pp_query : Send request to a paddle_packet cache to send Paddle packets with the given event id

paddle_packets : Receive paddle_packets from a paddle_packet cache

cmd_sender : Sending tof commands to comander thread. This is needed so that the event builder can request paddles from readout boards.