Function liftof_cc::threads::event_builder::event_builder

pub fn event_builder(
    m_trig_ev: &Receiver<MasterTriggerEvent>,
    ev_from_rb: &Receiver<RBEvent>,
    data_sink: &Sender<TofPacket>,
    run_id: u32,
    db_path: String,
    settings: TofEventBuilderSettings,
    thread_control: Arc<Mutex<ThreadControl>>
)

Events … assemble!

The event_builder collects all available event information, beginning with the MasterTriggerEvent defining the event id. It collects the requested number of RBEvents. The final product then will be a TofEvent

The event_builder is the heart of this software and crucial to all operations.

Arguments

m_trig_ev : Receive a MasterTriggerEvent over this channel. The event will be either build immediatly, or cached.

pp_query : Send request to a paddle_packet cache to send Paddle packets with the given event id

paddle_packets : Receive paddle_packets from a paddle_packet cache

cmd_sender : Sending tof commands to comander thread. This is needed so that the event builder can request paddles from readout boards.

settings : Configure the event builder