Function liftof_cc::threads::readoutboard_comm::readoutboard_communicator

pub fn readoutboard_communicator(
    ev_to_builder: &Sender<RBEvent>,
    tp_to_sink: Sender<TofPacket>,
    rb: &ReadoutBoard,
    runid: usize,
    print_packets: bool,
    run_analysis_engine: bool
)

Receive data from a readoutboard

In case of binary event data (“blob”) this can be analyzed here It is also possible to save the data directly.

In case of monitoring/other tof packets, those will be forwarded

Arguments:

pp_pusher :

reso_to_main :

tp_to_sink : Channel which should be connect to a (global) data sink. Packets which are of not event type (e.g. header/full binary data) will be forwarded to the sink.

write_rb_raw : Should readoutboard raw data be written to disk?

storage_savepath :

events_per_file :

rb :

runid : Current assigned runid. Will be used in the filenames of saved readoutboard raw data.

print_packets :